+++ src/main/java/de/maxhenkel/easyvillagers/integration/waila/HUDHandlerBreeder.java
@@ -4,7 +4,6 @@ import de.maxhenkel.easyvillagers.blocks.tileentity.BreederTileentity;
 import mcp.mobius.waila.api.IComponentProvider;
 import mcp.mobius.waila.api.IDataAccessor;
 import mcp.mobius.waila.api.IPluginConfig;
-import net.minecraft.util.text.IFormattableTextComponent;
 import net.minecraft.util.text.ITextComponent;
 
 import java.util.List;
@@ -20,12 +19,12 @@ public class HUDHandlerBreeder implements IComponentProvider {
         }
         BreederTileentity tileEntity = (BreederTileentity) accessor.getTileEntity();
 
-        IFormattableTextComponent villager1 = PluginEasyVillagers.getVillager(tileEntity.getVillagerEntity1());
+        ITextComponent villager1 = PluginEasyVillagers.getVillager(tileEntity.getVillagerEntity1());
         if (villager1 != null) {
             tooltip.add(villager1);
         }
 
-        IFormattableTextComponent villager2 = PluginEasyVillagers.getVillager(tileEntity.getVillagerEntity2());
+        ITextComponent villager2 = PluginEasyVillagers.getVillager(tileEntity.getVillagerEntity2());
         if (villager2 != null) {
             tooltip.add(villager2);
         }