+++ changelog.md
@@ -1,2 +1,3 @@
-- Fixed crash with XNet
-- Fixed exception when playing sound
+- Villagers can now be named in the anvil
+- Villager items show the villagers custom name
+- Fixed HWYLA tooltip not showing custom villager names